backend/registration-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
pyjwt
httpx
email-validator